+++ tfjs-core/src/backends/backend_util.ts
@@ -31,7 +31,8 @@ export * from '../ops/axis_util';
 export * from '../ops/broadcast_util';
 export * from '../ops/concat_util';
 export * from '../ops/conv_util';
-export {Activation, FusedConv2DConfig} from '../ops/fused_util';
+export * from '../ops/fused_util';
+export * from '../ops/fused/types';
 export * from '../ops/reduce_util';
 
 export {BackendValues, TypedArray, upcastType, PixelData} from '../types';
